refactor(integration-tests): share OOXML helpers in standard-terms-docx
Both employment-template-spacing.test.ts and
employment-template-style-spacing.test.ts duplicated DOCX zip loading,
xmldom traversal, namespace handling, paragraph text extraction, clause-
heading detection, and Standard Terms walking — only the assertions
differed. Extract a shared integration-tests/helpers/standard-terms-docx.ts
module exporting:

- W_NS, EMPLOYMENT_TEMPLATE_CASES
- loadXmlPart / loadDocumentXml
- directChild / wordAttr (depth-restricted, namespace-aware)
- paragraphText / paragraphStyle / paragraphSpacing
- isClauseHeadingParagraph / isDefinitionItemParagraph
- standardTermsParagraphs (body walker)

Both spacing tests collapse to ~50 lines from ~140; future OOXML drift
checks now reuse the same primitives.
